Remove 'tensorflow' pip package as a dependency in the setup.py file.
Top-level __init__.py file now checks that it can import TensorFlow (regardless of the pip package from which it comes from) and checks the TensorFlow version. If it can't import TensorFlow or if the installed version is too old, it prints an informative error message.

Reason:
Users who want to use GPUs must use the 'tensorflow-gpu' pip package.
PiperOrigin-RevId: 267359454