Move HolidayCalendar V1 and V2 under a common interface.
HolidayCalendar renamed to HolidayCalendarV1.
HolidayCalendar is now the interface which both V1 and V2
implement.
The calendars can now be created with dates.create_holiday_calendar,
which picks an appropriate impl; HolidayCalendarV1/2 are not exposed.
Thus, users don't have to know about two impls and their differences.

Also, merged the tests for V1 and V2.

PiperOrigin-RevId: 308796195